+++ core/components/modutilities/classes/modutilitiesrest.class.php
@@ -261,8 +261,15 @@
 
 				$this->log->set('input', $tt ?: NULL);
 				//сбор данных о подключении
+				$client = @$_SERVER['HTTP_CLIENT_IP'];
+				$forward = @$_SERVER['HTTP_X_FORWARDED_FOR'];
+				$remote = @$_SERVER['REMOTE_ADDR'];
+
+				if (filter_var($client, FILTER_VALIDATE_IP)) $ip = $client;
+				elseif (filter_var($forward, FILTER_VALIDATE_IP)) $ip = $forward;
+				else $ip = $remote;
 				$scriptProperties['user'] = [
-					'ip' => $this->util->getIP(),
+					'ip' => $ip,
 					'modxId' => $this->userId ?: 0,
 					'httpMethod' => @$_SERVER['REQUEST_METHOD'],
 				];